Move Auth0 user methods to Auth0UserTrait
The Auth0User class was using only a component of the Laravel
user model, \Illuminate\Contracts\Auth\Authenticatable. This left
out a lot of additional functionality.
